Move radiation variables
